Add LinkingManager
