+++ libsignal-service/src/provisioning/manager.rs
@@ -1,15 +1,27 @@
+use futures::{channel::mpsc::Sender, pin_mut, SinkExt, StreamExt};
 use phonenumber::PhoneNumber;
 use serde::{Deserialize, Serialize};
+use url::Url;
+
+use super::{
+    pipe::{ProvisioningPipe, ProvisioningStep},
+    ProvisioningError,
+};
 
 pub use crate::proto::{
     ProvisionEnvelope, ProvisionMessage, ProvisioningVersion,
 };
 
+use libsignal_protocol::{
+    generate_registration_id,
+    keys::{PrivateKey, PublicKey},
+    Context,
+};
+
 use crate::{
-    configuration::{Endpoint, ServiceConfiguration},
+    configuration::{Endpoint, ServiceConfiguration, SignalingKey},
     messagepipe::ServiceCredentials,
-    prelude::PushService,
-    push_service::{DeviceCapabilities, DeviceId, ServiceError},
+    push_service::{DeviceCapabilities, DeviceId, PushService, ServiceError},
     utils::{serde_base64, serde_optional_base64},
     USER_AGENT,
 };
@@ -108,6 +120,18 @@ pub enum VoiceVerificationCodeResponse {
 pub struct ProvisioningManager<P: PushService> {
     push_service: P,
     phone_number: PhoneNumber,
+}
+pub enum SecondaryDeviceProvisioning {
+    Url(Url),
+    NewDeviceRegistration {
+        phone_number: phonenumber::PhoneNumber,
+        device_id: DeviceId,
+        registration_id: u32,
+        uuid: String,
+        private_key: PrivateKey,
+        public_key: PublicKey,
+        profile_key: Vec<u8>,
+    },
 }
 
 impl<P: PushService> ProvisioningManager<P> {
@@ -241,5 +265,144 @@ impl<P: PushService> ProvisioningManager<P> {
         } else {
             format!("/v1/accounts/{}/code/{}", msg_type, phone_number)
         }
+    }
+}
+
+#[derive(Clone)]
+pub struct LinkingManager<P: PushService> {
+    push_service: P,
+    cfg: ServiceConfiguration,
+    password: String,
+}
+
+impl<P: PushService> LinkingManager<P> {
+    pub fn new(
+        cfg: impl Into<ServiceConfiguration> + Clone,
+        password: String,
+    ) -> Self {
+        Self {
+            password: password.clone(),
+            cfg: cfg.clone().into(),
+            push_service: P::new(cfg, None, USER_AGENT),
+        }
+    }
+
+    pub async fn provision_secondary_device(
+        &mut self,
+        ctx: &Context,
+        signaling_key: SignalingKey,
+        device_name: &str,
+        mut tx: Sender<SecondaryDeviceProvisioning>,
+    ) -> Result<(), ProvisioningError> {
+        // open a websocket without authentication, to receive a tsurl://
+        let (ws, stream) = self
+            .push_service
+            .ws("/v1/websocket/provisioning/", None)
+            .await?;
+
+        let registration_id = generate_registration_id(&ctx, 0)?;
+
+        let provisioning_pipe =
+            ProvisioningPipe::from_socket(ws, stream, &ctx)?;
+        let provision_stream = provisioning_pipe.stream();
+        pin_mut!(provision_stream);
+        while let Some(step) = provision_stream.next().await {
+            match step {
+                Ok(ProvisioningStep::Url(url)) => {
+                    tx.send(SecondaryDeviceProvisioning::Url(url))
+                        .await
+                        .expect("failed to send provisioning Url in channel");
+                }
+                Ok(ProvisioningStep::Message(message)) => {
+                    let uuid =
+                        message.uuid.ok_or(ProvisioningError::InvalidData {
+                            reason: "missing client UUID".into(),
+                        })?;
+
+                    let public_key = PublicKey::decode_point(
+                        &ctx,
+                        &message.identity_key_public.ok_or(
+                            ProvisioningError::InvalidData {
+                                reason: "missing public key".into(),
+                            },
+                        )?,
+                    )?;
+
+                    let private_key = PrivateKey::decode_point(
+                        &ctx,
+                        &message.identity_key_private.ok_or(
+                            ProvisioningError::InvalidData {
+                                reason: "missing public key".into(),
+                            },
+                        )?,
+                    )?;
+
+                    let profile_key = message.profile_key.ok_or(
+                        ProvisioningError::InvalidData {
+                            reason: "missing profile key".into(),
+                        },
+                    )?;
+
+                    let phone_number = message.number.ok_or(
+                        ProvisioningError::InvalidData {
+                            reason: "missing phone number".into(),
+                        },
+                    )?;
+
+                    let phone_number = phonenumber::parse(None, phone_number)
+                        .map_err(|e| {
+                        ProvisioningError::InvalidData {
+                            reason: format!("invalid phone number ({})", e),
+                        }
+                    })?;
+
+                    let mut provisioning_manager: ProvisioningManager<P> =
+                        ProvisioningManager::new(
+                            self.cfg.clone(),
+                            phone_number.clone(),
+                            self.password.to_string(),
+                        );
+
+                    let device_id = provisioning_manager
+                        .confirm_device(
+                            message
+                                .provisioning_code
+                                .ok_or(ProvisioningError::InvalidData {
+                                    reason: "no provisioning confirmation code"
+                                        .into(),
+                                })?
+                                .parse()
+                                .unwrap(),
+                            ConfirmDeviceMessage {
+                                signaling_key: signaling_key.to_vec(),
+                                supports_sms: false,
+                                fetches_messages: true,
+                                registration_id,
+                                name: device_name.to_string(),
+                            },
+                        )
+                        .await?;
+
+                    tx.send(
+                        SecondaryDeviceProvisioning::NewDeviceRegistration {
+                            phone_number,
+                            device_id,
+                            registration_id,
+                            uuid,
+                            private_key,
+                            public_key,
+                            profile_key,
+                        },
+                    )
+                    .await
+                    .expect(
+                        "failed to send provisioning message in rx channel",
+                    );
+                }
+                Err(e) => return Err(e.into()),
+            }
+        }
+
+        Ok(())
     }
 }
